Ensure package.json is writable in *.vsix
VSC requires the manifest file unpacked from the .vsix archive to be
marked as writable (see [1]). Since generating writable artifacts is
difficult in some build systems (such as Bazel), it would be helpful if
vsce itself could ensure that the file mode of the manifest is correct.

With this commit, the correct file mode on `package.json` is set
automatically by the packaging process via a new `Processor` and yazl's
`mode` argument.

[1]: https://github.com/microsoft/vscode/blob/88144f6d31718288c7a2c6fb741e0646d40804cf/src/vs/platform/extensionManagement/node/extensionsScanner.ts#L144