Make the [Content_Types].xml file deterministic
The order of the extension + mime types listed in the
[Content_Types].xml file is dependent on the order files are listed in
the files array, which can vary depending on the underlying filesystem's
traversal order. This can lead to non-reproducible .vsix files.

This sorts the contentTypes array to ensure the order is always
consistent and reproducible, regardless of the order of the files array,
ensuring reproducible .vsix files.