package: collect dependencies from parent dirs in a monorepo
The current packaging VSIX packaging strategy assumes that dependencies
are set up in a standalone directory. However, if a vscode package were
to exist in a `yarn` monorepo (e.g. one of the various large internal
monorepos at Microsoft), it would not be able to package.

This change adds support to that `yarn` wrapper for pulling dependencies
from the workspace's root `node_modules` directory, if this package is
run inside a monorepo.